services/user-service/requirements.txt
# Solace-AI User Service Dependencies
# Phase 9 - Production-ready dependencies

# Core Framework
pydantic==2.12.5
pydantic-settings==2.12.0

# Logging
structlog==25.5.0

# Security & Authentication
argon2-cffi==23.1.0          # Modern password hashing (Argon2id)
bcrypt==4.3.0                 # Legacy password support
PyJWT[crypto]==2.10.1         # JWT authentication
cryptography==44.0.0          # Encryption primitives
email-validator==2.2.0        # Production email validation

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
pytest-mock==3.15.1
pytest-timeout==2.4.0
faker==37.4.2
